apps/stt_stream/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
faster-whisper>=1.0.0
numpy>=1.24.0
nats-py>=2.9.0
websockets>=12.0
webrtcvad-wheels>=2.0.10
